Move view to correct directory
