+++ src/platform/web/ui/login/CompleteSSOView.js
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import {TemplateView} from "../platform/web/ui/general/TemplateView.js";
+import {TemplateView} from "../general/TemplateView.js";
 
 export class CompleteSSOView extends TemplateView {
     render(t) {
